+++ support/support.go
@@ -537,6 +537,8 @@ func (i *Info) printOSInfo() {
 			format(12, true, "Container", "Yes (Podman)")
 		case "lxc":
 			format(12, true, "Container", "Yes (LXC)")
+		case "containerd":
+			format(12, true, "Container", "Yes (containerd)")
 		case "yandex":
 			format(12, true, "Container", "Yes (Yandex Serverless)")
 		}