+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **`[system/container]`** Added [`containerd`](https://containerd.io) support
 - **`[system/container]`** Added [`K8s`](https://kubernetes.io) support
+- **`[support]`** Added `containerd` support
 
 ### [13.20.4](https://kaos.sh/ek/13.20.4)