+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### [13.20.5](https://kaos.sh/ek/13.20.5)
+
+- **`[system/container]`** Added [`containerd`](https://containerd.io) support
+- **`[system/container]`** Added [`K8s`](https://kubernetes.io) support
+- **`[support]`** Added `containerd` support
+
 ### [13.20.4](https://kaos.sh/ek/13.20.4)
 
 - **`[req]`** Added new helpers for working with `Query`